third_party/WebKit/LayoutTests/platform/linux/virtual/outofblink-cors-http/http/tests/security/video-cross-origin-via-dom-expected.txt

CONSOLE ERROR: Uncaught (in promise) NotSupportedError: Failed to load because no supported source was found.
This is a testharness.js-based test.
FAIL Untitled assert_unreached: Reached unreachable code
Harness: the test ran to completion.
